requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow==2.16.1
gdown
gradio
python-multipart
